# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "wavynote",
            "email": "chelseafandev@gmail.com"
        }
    },
    "paths": {
        "/box/conversation": {
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "특정 대화방 삭제",
                "tags": [
                    "받은노트 페이지"
                ],
                "summary": "특정 대화방 삭제",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "uid",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "conversation id",
                        "name": "cid",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/restapi.DefaultResponse"
                        }
                    },
                    "400": {
                        "description": "요청에 포함된 파라미터 값이 잘못된 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response400"
                        }
                    },
                    "401": {
                        "description": "인증에 실패한 경우이며, 실패 사유가 전달됩니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response401"
                        }
                    },
                    "404": {
                        "description": "요청한 리소스가 서버에 존재하지 않는 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response404"
                        }
                    },
                    "500": {
                        "description": "요청을 처리하는 과정에서 서버에 문제가 발생한 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response500"
                        }
                    }
                }
            }
        },
        "/box/conversationlist": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "노트를 주고 받는 대화방 목록 조회(최대 3개)",
                "tags": [
                    "받은노트 페이지"
                ],
                "summary": "노트를 주고 받는 대화방 목록 조회(최대 3개)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "조회된 대화방 목록 정보",
                        "schema": {
                            "$ref": "#/definitions/restapi.ConversationListResponse"
                        }
                    },
                    "400": {
                        "description": "요청에 포함된 파라미터 값이 잘못된 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response400"
                        }
                    },
                    "401": {
                        "description": "인증에 실패한 경우이며, 실패 사유가 전달됩니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response401"
                        }
                    },
                    "404": {
                        "description": "요청한 리소스가 서버에 존재하지 않는 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response404"
                        }
                    },
                    "500": {
                        "description": "요청을 처리하는 과정에서 서버에 문제가 발생한 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response500"
                        }
                    }
                }
            }
        },
        "/box/notelist": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "특정 친구와 주고받은 노트 목록",
                "tags": [
                    "받은노트 페이지"
                ],
                "summary": "특정 친구와 주고받은 노트 목록",
                "parameters": [
                    {
                        "type": "string",
                        "description": "conversation id",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/restapi.DefaultResponse"
                        }
                    },
                    "400": {
                        "description": "요청에 포함된 파라미터 값이 잘못된 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response400"
                        }
                    },
                    "401": {
                        "description": "인증에 실패한 경우이며, 실패 사유가 전달됩니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response401"
                        }
                    },
                    "404": {
                        "description": "요청한 리소스가 서버에 존재하지 않는 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response404"
                        }
                    },
                    "500": {
                        "description": "요청을 처리하는 과정에서 서버에 문제가 발생한 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response500"
                        }
                    }
                }
            }
        },
        "/box/show": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "특정 노트 조회",
                "tags": [
                    "받은노트 페이지"
                ],
                "summary": "특정 노트 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "conversation id",
                        "name": "cid",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "note id",
                        "name": "nid",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "uid",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "조회한 노트 정보",
                        "schema": {
                            "$ref": "#/definitions/restapi.NoteInfo"
                        }
                    },
                    "400": {
                        "description": "요청에 포함된 파라미터 값이 잘못된 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response400"
                        }
                    },
                    "401": {
                        "description": "인증에 실패한 경우이며, 실패 사유가 전달됩니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response401"
                        }
                    },
                    "404": {
                        "description": "요청한 리소스가 서버에 존재하지 않는 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response404"
                        }
                    },
                    "500": {
                        "description": "요청을 처리하는 과정에서 서버에 문제가 발생한 경우입니다",
                        "schema": {
                            "$ref": "#/definitions/restapi.Response500"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "restapi.ConversationInfo": {
            "type": "object",
            "properties": {
                "conversation_id": {
                    "description": "대화방 고유의 id 값",
                    "type": "string",
                    "example": "1afc571d-61bf-4cef-95ce-ab791f999297"
                },
                "new_note": {
                    "description": "대화방에 읽지않은 노트가 존재하는지 여부",
                    "type": "boolean",
                    "example": true
                },
                "note_count": {
                    "description": "대화방에 존재하는 노트의 총 개수",
                    "type": "integer",
                    "example": 20
                },
                "opp_nickname": {
                    "description": "대화 상대의 별명",
                    "type": "string",
                    "example": "somebody"
                }
            }
        },
        "restapi.ConversationListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/restapi.ConversationInfo"
                    }
                }
            }
        },
        "restapi.DefaultResponse": {
            "type": "object",
            "properties": {
                "msg": {
                    "description": "실패 시 반환하는 에러 메시지",
                    "type": "string"
                },
                "result": {
                    "description": "요청에 대한 처리 성공/실패 여부",
                    "type": "string",
                    "example": "true"
                }
            }
        },
...
        "restapi.NoteInfo": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "노트의 본문 내용",
                    "type": "string",
                    "example": "This is the main text of my first wavy note."
                },
                "conversation_id": {
                    "description": "대화방의 고유 id 값",
                    "type": "string",
                    "example": ""
                },
                "folder_id": {
                    "description": "노트가 포함되어 있는 폴더의 고유 id 값",
                    "type": "string",
                    "example": "980e71ba-0395-49aa-833e-3ebc76b3ec88"
                },
                "from_id": {
                    "description": "노트 작성자(또는 송신자)의 id",
                    "type": "string",
                    "example": "wavynoteadmin@gmail.com"
                },
                "keywords": {
                    "description": "노트의 키워드",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "b0d88d67-01fd-47f8-b426-6ca0657d0f6e"
                    ]
                },
                "note_id": {
                    "description": "노트의 고유 id 값",
                    "type": "string",
                    "example": "09d05df1-2958-4a3d-b910-3b4fb079327b"
                },
                "save_at": {
                    "description": "노트를 저장한 마지막 날짜 및 시간 정보",
                    "type": "string",
                    "example": "2023-11-01 21:00:00"
                },
                "send_at": {
                    "description": "노트를 송신한 날짜 및 시간 정보",
                    "type": "string",
                    "example": "2023-11-01 23:20:12"
                },
                "title": {
                    "description": "노트의 제목",
                    "type": "string",
                    "example": "my first note"
                },
                "to_id": {
                    "description": "노트 수신자의 id",
                    "type": "string",
                    "example": "somebody@naver.com"
                }
            }
        },
...
        "restapi.Response400": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "msg": {
                    "type": "string",
                    "example": "Bad request"
                }
            }
        },
        "restapi.Response401": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 401
                },
                "msg": {
                    "type": "string",
                    "example": "Fail to parse Authorization header"
                }
            }
        },
        "restapi.Response404": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 404
                },
                "msg": {
                    "type": "string",
                    "example": "Not exist the target resource in our server"
                }
            }
        },
        "restapi.Response500": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "msg": {
                    "type": "string",
                    "example": "Internal server error"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}